+++ CHANGELOG.md
@@ -5,6 +5,7 @@ ___
 
 Work in progress, to be added to next release notes.
 
+* Fix bug with StackView Transitioner where active screen content could remain hidden and prevent setState from being called while unmounting. ([#101](https://github.com/Roblox/roact-navigation/pull/101))
 * Allow for non-overlay screens to still use a transparent background and transparent background on non-overlay screens, change ClipsDescendants to false for navigator views and fix "Listener disconnected twice" error ([#97](https://github.com/Roblox/roact-navigation/pull/97))
 * Add `RoactNavigation.None` to allow removal of params ([#90](https://github.com/Roblox/roact-navigation/pull/90))
 * Add second optional parameter to `createAppContainer` to provide a LinkingProtocol object. This will be used when connecting the universal app to use deep-linking ([#83](https://github.com/Roblox/roact-navigation/pull/83))